refactor: 기존 carFixture 의미에 맞게 상수 이름으로 변경
