+++ src/test/java/racingcar/domain/CarTest.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static racingcar.domain.fixture.CarFixture.CAR_NAME;
+import static racingcar.domain.fixture.ConstantsFixture.CAR_NAME;
 
 class CarTest {
