+++ src/test/java/racingcar/domain/fixture/ConstantsFixture.java
@@ -1,6 +1,6 @@
 package racingcar.domain.fixture;
 
-public class CarFixture {
+public class ConstantsFixture {
 
     public static final String CAR_NAMES = "pobi,crong,honux";
 
@@ -15,5 +15,4 @@ public class CarFixture {
     public static final String CAR_NAME_ERROR_MESSAGE = "자동차 이름을 다시 확인해 주세요.";
 
     public static final String TRY_COUNT_ERROR_MESSAGE = "시도 횟수는 0 보다 커야 합니다.";
-
 }
